+++ CHANGELOG.md
@@ -5,6 +5,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v2.0.5] - 2019-12-15
+### Added
+- Update requirements to support python 3.8
+
+## [v2.0.4] - 2019-11-19
+### Fixed
+- Fix static app resolving
+
+## [v2.0.3] - 2019-09-20
+### Fixed
+- Fix template conflicts
+
+## [v2.0.2] - 2019-09-20
+### Fixed
+- Fix template conflicts
+
+## [v2.0.1] - 2019-09-20
+### Fixed
+- Fix template import
 
 ## [v2.0.0] - 2019-09-19
 ### Changed
@@ -212,7 +231,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Conception!
 
-[Unreleased]: https://github.com/taoufik07/responder/compare/v2.0.0..HEAD
+[Unreleased]: https://github.com/taoufik07/responder/compare/v2.0.5..HEAD
+[v2.0.5]: https://github.com/taoufik07/responder/compare/v2.0.4..v2.0.5
+[v2.0.4]: https://github.com/taoufik07/responder/compare/v2.0.3..v2.0.4
+[v2.0.3]: https://github.com/taoufik07/responder/compare/v2.0.2..v2.0.3
+[v2.0.2]: https://github.com/taoufik07/responder/compare/v2.0.1..v2.0.2
+[v2.0.1]: https://github.com/taoufik07/responder/compare/v2.0.0..v2.0.1
 [v2.0.0]: https://github.com/taoufik07/responder/compare/v1.3.2..v2.0.0
 [v1.3.2]: https://github.com/taoufik07/responder/compare/v1.3.1..v1.3.2
 [v1.3.1]: https://github.com/taoufik07/responder/compare/v1.3.0..v1.3.1